template header files contain function codes as well and removed redundant subfolders
